realtime/trade-collectors-secondary/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
redis[hiredis]>=5.0.0
python-dotenv>=1.0.0
